fields-classes sorted by modules
