+++ src/fields/__init__.py
@@ -1,19 +1,17 @@
-from .fields import (
-    ArrayField,
-    BaseField,
-    BooleanField,
-    DictField,
-    FloatField,
-    IntegerEnumField,
-    IntegerField,
-    OneOfField,
-    ReferenceField,
-    StringEnumField,
-    StringField,
-    UnionField,
-    get_enum_field_from_dict,
-    get_field_from_dict,
-)
+from .array import ArrayField
+from .base import BaseField
+from .boolean import BooleanField
+from .dict import DictField
+from .fields import get_enum_field_from_dict, get_field_from_dict
+from .float import FloatField
+from .integer import IntegerField
+from .integer_enum import IntegerEnumField
+from .one_of import OneOfField
+from .pattern import PatternField
+from .reference import ReferenceField
+from .string import StringField
+from .string_enum import StringEnumField
+from .union import UnionField
 
 __all__ = [
     "BaseField",
@@ -26,6 +24,7 @@ __all__ = [
     "ReferenceField",
     "UnionField",
     "OneOfField",
+    "PatternField",
     "StringEnumField",
     "IntegerEnumField",
     "get_field_from_dict",
