+++ src/fields/dict.py
@@ -1,0 +1,9 @@
+from .base import BaseField
+
+
+class DictField(BaseField):
+    type: str
+
+    @property
+    def __typehint__(self) -> str:
+        return "dict"
