+++ src/fields/fields.py
@@ -1,323 +1,18 @@
-from __future__ import annotations
+from src.strings import get_reference
 
-from typing import Literal, Optional, Union
-
-from msgspec import Struct
-
-from src.strings import get_reference, is_valid_name, to_python_types, validate_field
-
-
-class BaseField(Struct):
-    type: Union[str, list]
-    name: str
-    required: bool = False
-    """If required is not defined, it is assumed to be false."""
-    description: Optional[str] = None
-
-    @property
-    def __typehint__(self) -> str:
-        raise NotImplementedError
-
-    def _get_required_field_class(self) -> str:
-        if not self.required:
-            raise ValueError("Field is not required")
-        name_is_valid = is_valid_name(self.name)
-        if not name_is_valid:
-            name = validate_field(self.name)
-            return (
-                f"    {name}: {self.__typehint__} = pydantic.Field(\n"
-                f'        alias="{self.name}"\n'
-                f"    )\n"
-            )
-        return f"    {self.name}: {self.__typehint__}\n"
-
-    def _get_optional_field_class(self) -> str:
-        if self.required:
-            raise ValueError("Field is required")
-        name_is_valid = is_valid_name(self.name)
-        if not name_is_valid:
-            name = validate_field(self.name)
-            return (
-                f"    {name}: typing.Optional[{self.__typehint__}] = pydantic.Field(\n"
-                f'        default=None, alias="{self.name}"\n'
-                f"    )\n"
-            )
-        return f"    {self.name}: typing.Optional[{self.__typehint__}] = None\n"
-
-    def _get_description(self) -> str:
-        if self.description is None:
-            return ""
-        return f'    """{self.description}"""\n'
-
-    def to_field_class(self):
-        if self.required:
-            string = self._get_required_field_class()
-        else:
-            string = self._get_optional_field_class()
-
-        string += self._get_description()
-        return string
-
-
-class ReferenceField(BaseField):
-    type: str = "reference"
-    """
-    This field is missing from original schema for this property,
-    but it is required for parsing.
-    """
-    reference: str
-    default: Optional[str] = None
-
-    @property
-    def __typehint__(self) -> str:
-        return self.reference
-
-    def _get_default_field_class(self) -> str:
-        if self.default is None:
-            raise ValueError("Default value is not defined")
-        name_is_valid = is_valid_name(self.name)
-        # If reference to have default value, then this value is the field of enum
-        default_value = f"{self.__typehint__}.{self.default.upper()}"
-        if not name_is_valid:
-            name = validate_field(self.name)
-            return (
-                f"    {name}: {self.__typehint__} = pydantic.Field(\n"
-                f'        default={default_value}, alias="{self.name}"\n'
-                f"    )\n"
-            )
-        return f"    {self.name}: {self.__typehint__} = {default_value}\n"
-
-    def to_field_class(self):
-        if self.default is not None:
-            string = self._get_default_field_class()
-        elif self.required:
-            string = self._get_required_field_class()
-        else:
-            string = self._get_optional_field_class()
-
-        string += self._get_description()
-        return string
-
-
-class StringField(BaseField):
-    type: str
-    maxLength: Optional[int] = None
-    format: Optional[Literal["uri"]] = None
-
-    @property
-    def __typehint__(self) -> str:
-        return "str"
-
-
-class IntegerField(BaseField):
-    type: str
-    default: Optional[int] = None
-    minimum: Optional[int] = None
-    maximum: Optional[int] = None
-    entity: Optional[Literal["owner"]] = None
-    format: Optional[Literal["int64"]] = None
-
-    @property
-    def __typehint__(self) -> str:
-        return "int"
-
-    def _get_default_field_class(self) -> str:
-        if self.default is None:
-            raise ValueError("Default value is not defined")
-        name_is_valid = is_valid_name(self.name)
-        if not name_is_valid:
-            name = validate_field(self.name)
-            return (
-                f"    {name}: {self.__typehint__} = pydantic.Field(\n"
-                f'        default={self.default}, alias="{self.name}"\n'
-                f"    )\n"
-            )
-        return f"    {self.name}: {self.__typehint__} = {self.default}\n"
-
-    def to_field_class(self):
-        if self.default is not None:
-            string = self._get_default_field_class()
-        elif self.required:
-            string = self._get_required_field_class()
-        else:
-            string = self._get_optional_field_class()
-
-        string += self._get_description()
-        return string
-
-
-class FloatField(BaseField):
-    type: str
-    minimum: Optional[Union[int, float]] = None
-    maximum: Optional[int] = None
-
-    @property
-    def __typehint__(self) -> str:
-        return "float"
-
-
-class BooleanField(BaseField):
-    type: str
-    default: Optional[bool] = None
-
-    @property
-    def __typehint__(self) -> str:
-        return "bool"
-
-    def _get_default_field_class(self) -> str:
-        if self.default is None:
-            raise ValueError("Default value is not defined")
-        name_is_valid = is_valid_name(self.name)
-        typehint = self.__typehint__
-        if not name_is_valid:
-            name = validate_field(self.name)
-            return (
-                f"    {name}: {typehint} = pydantic.Field(\n"
-                f'        default={self.default}, alias="{self.name}"\n'
-                f"    )\n"
-            )
-        return f"    {self.name}: {typehint} = {self.default}\n"
-
-    def to_field_class(self):
-        if self.default is not None:
-            string = self._get_default_field_class()
-        elif self.required:
-            string = self._get_required_field_class()
-        else:
-            string = self._get_optional_field_class()
-
-        string += self._get_description()
-        return string
-
-
-class DictField(BaseField):
-    type: str
-
-    @property
-    def __typehint__(self) -> str:
-        return "dict"
-
-
-class ArrayField(BaseField):
-    type: str
-    items: BaseField
-
-    @property
-    def __typehint__(self) -> str:
-        return f"list[{self.items.__typehint__}]"
-
-    def _get_description(self) -> str:
-        if self.items.description is not None:
-            return f'    """{self.items.description}"""\n'
-        elif self.description is not None:
-            return f'    """{self.description}"""\n'
-        return ""
-
-    def to_field_class(self):
-        if self.required:
-            string = self._get_required_field_class()
-        else:
-            string = self._get_optional_field_class()
-
-        string += self._get_description()
-        return string
-
-
-class UnionField(BaseField):
-    type: list[str]
-
-    @property
-    def __typehint__(self) -> str:
-        python_types = to_python_types(self.type)
-        types = ", ".join(python_types)
-        return f"typing.Union[{types}]"
-
-
-class OneOfField(BaseField):
-    type: str = "oneOf"
-    """
-    This field is missing from original schema for this property,
-    but it is required for parsing.
-    """
-    oneOf: list[BaseField]
-
-    @property
-    def __typehint__(self) -> str:
-        typehints = [field.__typehint__ for field in self.oneOf]
-        types = ", ".join(typehints)
-        return f"typing.Union[{types}]"
-
-
-class StringEnumField(StringField):
-    __typehint__: str
-
-    type: str
-    enum: list[str]
-    enumNames: Optional[list[str]] = None
-
-
-class IntegerEnumField(IntegerField):
-    __typehint__: str
-
-    type: str
-    enum: list[int]
-    enumNames: list[str]
-
-    def _get_default_enum(self) -> str:
-        """
-        Integer enum can have default value, for getting name of enum
-        we need to get value of enum.
-        :return: name of enum
-        """
-        for enum, enum_name in zip(self.enum, self.enumNames):
-            if enum == self.default:
-                return enum_name
-        raise ValueError("Default value is not defined")
-
-    def _get_default_field_class(self) -> str:
-        if self.default is None:
-            raise ValueError("Default value is not defined")
-        name_is_valid = is_valid_name(self.name)
-        default_value = self._get_default_enum()
-        default = f"{self.__typehint__}.{default_value.upper()}"
-        if not name_is_valid:
-            name = validate_field(self.name)
-            return (
-                f"    {name}: {self.__typehint__} = pydantic.Field(\n"
-                f'        default={default}, alias="{self.name}"\n'
-                f"    )\n"
-            )
-        return f"    {self.name}: {self.__typehint__} = {default}\n"
-
-
-class PatternField(BaseField):
-    type: str = "patternProperties"
-    """
-    This field is missing from original schema for this property,
-    but it is required for parsing.
-    """
-    patternProperties: dict[str, BaseField]
-    additionalProperties: bool
-
-    @property
-    def __typehint__(self) -> str:
-        typehints = [field.__typehint__ for field in self.patternProperties.values()]
-        types = ", ".join(typehints)
-        if len(typehints) == 1:
-            return f"dict[str, {types}]"
-        return f"dict[str, typing.Union[{types}]]"
-
-    def _get_description(self) -> str:
-        if self.description is not None:
-            string = f'    """\n    {self.description}\n'
-        else:
-            string = '    """\n'
-
-        string += "    Patterns of dict (as regexp) in the form of key-value:\n"
-        for pattern, field in self.patternProperties.items():
-            string += f"    {pattern}: {field.__typehint__}\n"
-        string += '    """\n'
-        return string
+from .array import ArrayField
+from .base import BaseField
+from .boolean import BooleanField
+from .dict import DictField
+from .float import FloatField
+from .integer import IntegerField
+from .integer_enum import IntegerEnumField
+from .one_of import OneOfField
+from .pattern import PatternField
+from .reference import ReferenceField
+from .string import StringField
+from .string_enum import StringEnumField
+from .union import UnionField
 
 
 def get_field_from_dict(item: dict, name: str) -> BaseField:
