+++ src/fields/float.py
@@ -1,0 +1,13 @@
+from typing import Optional, Union
+
+from .base import BaseField
+
+
+class FloatField(BaseField):
+    type: str
+    minimum: Optional[Union[int, float]] = None
+    maximum: Optional[int] = None
+
+    @property
+    def __typehint__(self) -> str:
+        return "float"
