+++ src/fields/string.py
@@ -1,0 +1,13 @@
+from typing import Literal, Optional
+
+from .base import BaseField
+
+
+class StringField(BaseField):
+    type: str
+    maxLength: Optional[int] = None
+    format: Optional[Literal["uri"]] = None
+
+    @property
+    def __typehint__(self) -> str:
+        return "str"
